## 4.3.1.0
* Bug Fixes
  * will avoid leak of local reference in Java/JNI VSI layer's seek implementation
  * corrected issues with strings found during translation to Japanese
  * migrated the name of the directory for the icons from the incorrect mil-std-2525b to mil-std-2525c
  * will resize the text widget if needed when the bloodhound is started
  * migrated over fix for GlArrow2 so that the association arrow is not render sub-terrain when no altitude/elevation is present
  * corrected issue with FAH being clipped incorrectly in areas greater than 90 degree longitude
  * corrected issue with countries using translation for All Chat Room
* Feature Additions
  * can turn off enhanced depth perception mode
  * can modify the styling of features (preliminary capability)
  * migrated to Label Manager

+++ takengine/mapengine/sdk/src/raster/mosaic/ATAKMosaicDatabase.cpp
@@ -114,7 +114,7 @@ namespace atakmap {
                 }
 
                 void insertRow(const std::string &path, const std::string &type, core::GeoPoint ul, core::GeoPoint ur, core::GeoPoint lr,
-                                       core::GeoPoint ll, double minGsd, double maxGsd, int width, int height) override
+                                       core::GeoPoint ll, double minGsd, double maxGsd, size_t width, size_t height) override
                 {
                     Lock lock(filter->getMutex());
                     filter->insertRow(path, type, ul, ur, lr, ll, minGsd, maxGsd, width, height);
@@ -436,7 +436,7 @@ namespace atakmap {
             }
 
             void ATAKMosaicDatabase::insertRow(const std::string &path, const std::string &type, core::GeoPoint ul, core::GeoPoint ur, core::GeoPoint lr,
-                                               core::GeoPoint ll, double minGsd, double maxGsd, int width, int height)
+                                               core::GeoPoint ll, double minGsd, double maxGsd, size_t width, size_t height)
             {
 
             }
